% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/IAEA-ENDF-file-reader-photonuclear-crosssections-001.R
\name{clanENDF}
\alias{clanENDF}
\title{clanENDF}
\usage{
clanENDF()
}
\description{
Package for reading IAEA nuclear data files.
This is a simple dummy function to ease access to the help index.

\preformatted{
 clanENDF()

 ?clanENDF # gives you an index of functions in package
  }

Main functions for computation of the density-effect correction are:
Simple read function for nuclear data from IAEA files in ENDF format.

Location of the IAEA 2019 ENDF-files. Could be here:

   \preformatted{
    pn.full <- paste(getwd(),"/data/",sep="")}

Or if the data are coming with the clanENDF package you may
be able to get it using:

   \preformatted{
   #' pn.full <- paste(.libPaths(),"/clanENDF","/data/",sep="")}

How to read the data

   \preformatted{
   df <- read.selected.IAEA.photonuclear.data(
   pn = pn.full,
   save.data.to.file=TRUE,
   save.fn="IAEA-photonuclear-2019.txt")}

Then call the demo function:
  \preformatted{
    IAEA.photonuclear.demo()}
}
