% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/IAEA-ENDF-file-reader-photonuclear-crosssections-001.R
\name{endf.number}
\alias{endf.number}
\title{endf.number}
\usage{
endf.number(x = "4.555", y = "-3")
}
\arguments{
\item{x}{= first part of number (e.g. x="4.555" if number is x="4.555-3").}

\item{y}{= last part of number (e.g. y="-3" if number is x="4.555-3").}
}
\value{
a float.
}
\description{
Helper function for ENDF file numbers.
}
\details{
# Convert ENDF data to a floating point number.
A raw ENDF number is written as 2.224000+6 - i.e. without any E.
This function adds the E (e.g. 2.224000E+6) and convert to a float.
Simple read function for nuclear data from IAEA files in ENDF format.
}
