% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/IAEA-ENDF-file-reader-photonuclear-crosssections-001.R
\name{IAEA.photonuclear.demo}
\alias{IAEA.photonuclear.demo}
\title{IAEA.photonuclear.demo}
\usage{
IAEA.photonuclear.demo()
}
\description{
Demonstration function
}
\details{
Simple read function for nuclear data from IAEA files in ENDF format.

Location of the IAEA 2019 ENDF-files. Could be here:
  pn.full <- paste(getwd(),"/data/",sep="")
How to read the data
   df <- read.selected.IAEA.photonuclear.data(
   pn = pn.full,
   save.data.to.file=TRUE,
   save.fn="IAEA-photonuclear-2019.txt")

Then call the demo function:
  IAEA.photonuclear.demo()
}
