% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/IAEA-ENDF-file-reader-photonuclear-crosssections-001.R
\name{IAEA.photonuclear.plot}
\alias{IAEA.photonuclear.plot}
\title{IAEA.photonuclear.plot}
\usage{
IAEA.photonuclear.plot(
  pn = "",
  fn = "IAEA-photonuclear-2019.txt",
  log.wanted = TRUE,
  MeV.min = 0,
  MeV.max = 20.1
)
}
\arguments{
\item{pn}{= path to folder with endf data files (ending with a /).}

\item{fn}{= file name with the data to read (e.g. "IAEA-photonuclear-2019.txt").}

\item{log.wanted}{= TRUE or FALSE.}

\item{MeV.min}{= lower energy interval in MeV.}

\item{MeV.max}{= upper energy interval in MeV.}
}
\value{
a float.
}
\description{
Read file with all data and make a trellis plot.
}
\details{
# Created: March 10, 2023
Revised: March 11, 2023
Revised: March 13, 2023
Name: Claus E. Andersen
Objective take the IAEA data from the IAEA in the file file fn,
and show the results.
}
