% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/IAEA-ENDF-file-reader-photonuclear-crosssections-001.R
\name{read.endf}
\alias{read.endf}
\title{read.endf}
\usage{
read.endf(
  pn = "",
  fn = "g_1-H-2_0128.endf",
  line.start = 91,
  line.stop = 100,
  Z = 1,
  A = 2,
  element = "H",
  isotope = "H-2",
  process = "",
  what = "crosssection for (gamma,n)"
)
}
\arguments{
\item{pn}{= path to folder with endf data files (ending with a /).}

\item{fn}{= file name (e.g. "g_1-H-2_0128.endf")}

\item{line.start}{=  = last line in the endf file that you need to extract.}

\item{Z}{= atomic number (meta data).}

\item{A}{= atomic mass (meta data).}

\item{element}{= element name (e.g. "H") (meta data).}

\item{isotope}{= isotope name (e.g. "H-2") (meta data).}

\item{process}{= name of process covered by cross section (meta data).}

\item{what}{= free text description (meta data).}
}
\value{
a data frame with data from endf file + meta data
}
\description{
Simple read function for nuclear data from IAEA files in ENDF format.
}
\details{
Simple read function for nuclear data from IAEA files in ENDF format.

The line.start and line.stop can be found by manual inspection
of the endf data file. The read.endf function can only read
the pure data (no meta data).

Meta data can be assigned to the data. This facilitates
data from different isotopes or processes etc.

#########################################################
 Main function read endf-files
 The user has to manually identify the part of the ENDF file which
 is of interest by given the start and stop lines.
 The user can assign meta data to the results.
########################################################
 Created: March 10, 2023
Revised: March 11, 2023
Name   : Claus E. Andersen
Motivation:
To read IAEA 2019 photonuclear data.
Input:
pn = path to where the ENDF file is
fn = file name of ENDF file
line.start = Where to start reading from the fine (line number)
line.stop  = Where to stop reading from the fine (line number)
Z = Atomic number (meta data to be added to the output)
A = Atomic mass (meta data to be added to the output)
element = Element name (meta data to be added to the output)
isotope = Isotope name (meta data to be added to the output)
process = Process name (meta data to be added to the output)
what    = Description (meta data to be added to the output)
Output:
A data frame like this:
    MeV      barn              file Z A element isotope process
1 2.224 0.0000000 g_1-H-2_0128.endf 1 2       H     H-2   total
2 2.300 0.0006581 g_1-H-2_0128.endf 1 2       H     H-2   total
3 2.600 0.0011830 g_1-H-2_0128.endf 1 2       H     H-2   total
4 3.000 0.0018280 g_1-H-2_0128.endf 1 2       H     H-2   total
5 3.500 0.0022800 g_1-H-2_0128.endf 1 2       H     H-2   total
6 4.000 0.0024540 g_1-H-2_0128.endf 1 2       H     H-2   total

                           what
1 cross section for (gamma,abs)
2 cross section for (gamma,abs)
3 cross section for (gamma,abs)
4 cross section for (gamma,abs)
5 cross section for (gamma,abs)
6 cross section for (gamma,abs)
}
